+++ witch_craft/src/core/consts.rs
