+++ witch_craft/src/core/mod.rs
@@ -2,3 +2,4 @@ pub mod consts;
 pub mod core;
 pub mod data;
 pub mod structs;
+pub mod types;
