+++ witch_craft/src/core/types.rs
@@ -1,0 +1,1 @@
+pub type Closure = Vec<(&'static str, fn(&[String]) -> i32)>;
