+++ witch_craft/src/modules/blackcat/mod.rs
@@ -1,1 +1,2 @@
+pub mod backend;
 pub mod blackcat;
