+++ witch_craft/src/modules/mod.rs
@@ -1,6 +1,6 @@
 pub mod binds;
 pub mod blackcat;
-pub mod core;
 pub mod llm;
 pub mod network;
-pub mod shell;
+pub mod osint;
+pub mod social;
