+++ witch_craft/src/modules/network/mod.rs
@@ -1,2 +1,4 @@
+pub mod ddos;
+pub mod map_network;
 pub mod network;
 pub mod structs;
