+++ witch_craft/src/modules/osint/mod.rs
@@ -1,0 +1,1 @@
+pub mod osint;
