--- witch_craft/src/modules/shell/mod.rs
@@ -1,1 +1,0 @@
-pub mod shell;
