+++ witch_craft/src/modules/social/mod.rs
@@ -1,0 +1,1 @@
+pub mod social;
