--- witch_craft/src/test/malware_list.txt
@@ -1,35 +1,0 @@
-################################################################
-# MalwareBazaar full malware samples dump (SHA256 hashes)      #
-# Last updated: 2024-08-20 18:30:44 UTC                        #
-#                                                              #
-# Terms Of Use: https://bazaar.abuse.ch/faq/#tos               #
-# For questions please contact bazaar [at] abuse.ch            #
-################################################################
-#
-# sha256_hash
-8d31c82108ecfafc5d22165857de868594b5e6529d0709c279d297b50ad60fd5
-2a7a97fe1b769f2b74ebd66c447708f5b5beb60bad5a53d05d7f428770ba2f62
-41d16cf08c360bd64f59a76480ab0560e8d466011075453e17182940d93b6f24
-a61dc154af80d7c67638c50ca91d567ba0f872562c1b6c616e58abdda3bc3544
-fdb65b07bb55b2b7c0a7515a1a56081dad716678649c5345df5de33186ec5679
-21381b405bbb2d1ac38f1d908e0dc8a399fb2401d2ed1c1a300a2144626f9add
-88f5038ef6b234a60a7d26fe211e14ad488a99a5926dd4f5c1ea48d4ea551eab
-b920dc19a2317f619a9d7af0935eb05b07442d2ae77f1482bd883a086a9c0513
-b3e2fe43f3024cc479415e745cd9826752debe4e8b208e5e5b7cc510723b787d
-0f8eac72a742519dfb28cf96c917b4e82532ea5c6332ed9df105ed5adcd2f421
-74adc4039c75425ad6a2b19b4ab6ac460d9a21f30ae71bb71408b8f530907c9b
-b84c59b20bddf6db3f6feec223dacc4b194350eca67f393df64529394c1a52b4
-7bc7edf2f2fafaa8457fb596cbbcdedafd23544d75e739e777b73790965df6bb
-773fe822889cd0d959c1a672143ca552cc18406ee4c7e2f16c25b3c672d0ed8f
-fbe42f49f702bf951e4c1f4ec953db5a142481790d93190087927382d91723a0
-f961dd0541b927a01e09a97c8d371716e8d8ab35b3d7293dbcd9e93ee335b114
-24307f7e934f287415e74aa439167fa9ea677676da94deee14e42d1816312934
-790efbfa7e44ecb1f724e6babb44429c6dbb1c0e46297e7252eb6bb4d56f9c35
-1528e4f9efb7c5d50510d8c9343e793920554df874b88306758c55e420d57dbe
-f346a3c9f8d6ad626f820f1d009ca3059e9fac6f260aa63a861fdb247f7b6526
-ecba82efb819da9c18bcfd877077ef1a6500525dd7de736b3939ef337cb948a1
-343392269e2048b507fb005e66cb97ee698a13f9a1990bddcb5f6ce4ed26f8e8
-78bc22e7ab1e73308787b284f3342510c75547cb854dec4da5ba5d523979a33c
-9f68e60790c6413be7ef8776ce90b064e67cf80505126de8af9de94ec50ad70a
-d89149126795ce5c323a155792b51d7c08c2dda36be28f559dfbdab43a4e538d
-test/malware_list.txt
